refactor(layout): move gnn/ to layers/gnn/ (T124.5.1)
