+++ layers/gnn/doc.go
